Resume Hosting the 834 Companion Guide (#9)
Revert the last change we made to temporarily not host
the CMS 834 companion guide. CMS said we could host
it, but also provided the link to the guide on their site.
We tried just linking to them, but that requires the person
downloading it to log into CMS. Then, it requires that you
have access to it. When I went to request access, it said I
didn't have access to do that! We figured that's a bad
user experience, so we're just going to go back to hosting
it.

Revert "Remove the CMS 834 guide and links (#6)"

This reverts commit f342d151be95b91bab371e5ed8e0254c6a99e11d.

Co-authored-by: Timothy McGrath <tmcgrath@hioscar.com>